+++ src/core/bs_version.h
@@ -11,7 +11,7 @@
 
 #define BS_VERSION_MAJOR 0
 #define BS_VERSION_MINOR 4
-#define BS_VERSION_PATCH 1
+#define BS_VERSION_PATCH 2
 
 #define BS_STR2(x) #x
 #define BS_STR(x)  BS_STR2(x)